Changed commands and added String command
In order to set the tone the documentation reveals, a new structure grouping commands or functions into types, I created the String command, which will allow me to use any subcommand (like the previous uppercase and lowercase) from that group to a string. This organizes the amount of commands the app will have into a more logical distribution.