Created functions related to copy and it's local file
After some digging and testing, I learned that the values aren't stored inside Mars, since it's a CLI tool. To come around this, I need to store the copy and paste information locally in the computer, in a file. That process of creating, getting the values and saving them were declared into fileops.